+++ src/biz/bokhorst/xprivacy/XContentProvider.java
@@ -220,7 +220,7 @@ public class XContentProvider extends XHook {
 						while (cursor.moveToNext()) {
 							int colPackage = cursor.getColumnIndex("package");
 							String packageName = (colPackage < 0 ? null : cursor.getString(colPackage));
-							if (packageName != null && XApplicationPackageManager.isPackageAllowed(packageName))
+							if (packageName != null && XPackageManager.isPackageAllowed(packageName))
 								copyColumns(cursor, result);
 						}
 						result.respond(cursor.getExtras());