+++ src/biz/bokhorst/xprivacy/XPackageManager.java
@@ -7,6 +7,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import android.os.Binder;
+import android.os.Build;
 import android.util.Log;
 import android.content.Context;
 import android.content.pm.ApplicationInfo;
@@ -16,40 +17,53 @@ import android.content.pm.ResolveInfo;
 
 import de.robv.android.xposed.XC_MethodHook.MethodHookParam;
 
-public class XApplicationPackageManager extends XHook {
+public class XPackageManager extends XHook {
 	private Methods mMethod;
+	private String mClassName;
 
-	private XApplicationPackageManager(Methods method, String restrictionName) {
+	private XPackageManager(Methods method, String restrictionName, String className) {
 		super(restrictionName, method.name(), null);
 		mMethod = method;
+		mClassName = className;
 	}
 
 	public String getClassName() {
-		return "android.app.ApplicationPackageManager";
+		return mClassName;
 	}
 
 	// @formatter:off
 
 	// public List<ApplicationInfo> getInstalledApplications(int flags)
 	// public List<PackageInfo> getInstalledPackages(int flags)
+	// public String[] getPackagesForUid(int uid)
+	// public List<PackageInfo> getPackagesHoldingPermissions(String[] permissions, int flags)
 	// public List<PackageInfo> getPreferredPackages(int flags)
 	// public List<ResolveInfo> queryBroadcastReceivers(Intent intent, int flags)
 	// public List<ProviderInfo> queryContentProviders(String processName, int uid, int flags)
 	// public List<ResolveInfo> queryIntentActivities(Intent intent, int flags)
 	// public List<ResolveInfo> queryIntentActivityOptions(ComponentName caller, Intent[] specifics, Intent intent, int flags)
+	// public List<ResolveInfo> queryIntentContentProviders(Intent intent, int flags)
 	// public List<ResolveInfo> queryIntentServices(Intent intent, int flags)
 	// frameworks/base/core/java/android/app/ApplicationPackageManager.java
 	
 	// @formatter:on
 
 	private enum Methods {
-		getInstalledApplications, getInstalledPackages, getPreferredPackages, queryBroadcastReceivers, queryContentProviders, queryIntentActivities, queryIntentActivityOptions, queryIntentServices
+		getInstalledApplications, getInstalledPackages, getPackagesForUid, getPackagesHoldingPermissions, getPreferredPackages, queryBroadcastReceivers, queryContentProviders, queryIntentActivities, queryIntentActivityOptions, queryIntentContentProviders, queryIntentServices
 	};
 
-	public static List<XHook> getInstances() {
+	public static List<XHook> getInstances(Object instance) {
+		String className = instance.getClass().getName();
 		List<XHook> listHook = new ArrayList<XHook>();
 		for (Methods am : Methods.values())
-			listHook.add(new XApplicationPackageManager(am, PrivacyManager.cSystem));
+			if (am == Methods.getPackagesHoldingPermissions) {
+				if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR2)
+					listHook.add(new XPackageManager(am, PrivacyManager.cSystem, className));
+			} else if (am == Methods.queryIntentContentProviders) {
+				if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT)
+					listHook.add(new XPackageManager(am, PrivacyManager.cSystem, className));
+			} else
+				listHook.add(new XPackageManager(am, PrivacyManager.cSystem, className));
 		return listHook;
 	}
 
@@ -64,11 +78,16 @@ public class XApplicationPackageManager extends XHook {
 		if (mMethod == Methods.getInstalledApplications) {
 			if (param.getResult() != null && isRestricted(param))
 				param.setResult(filterApplicationInfo((List<ApplicationInfo>) param.getResult()));
-		} else if (mMethod == Methods.getInstalledPackages || mMethod == Methods.getPreferredPackages) {
+		} else if (mMethod == Methods.getPackagesForUid) {
+			if (param.getResult() != null && isRestricted(param))
+				param.setResult(new String[0]);
+		} else if (mMethod == Methods.getInstalledPackages || mMethod == Methods.getPackagesHoldingPermissions
+				|| mMethod == Methods.getPreferredPackages) {
 			if (param.getResult() != null && isRestricted(param))
 				param.setResult(filterPackageInfo((List<PackageInfo>) param.getResult()));
 		} else if (mMethod == Methods.queryBroadcastReceivers || mMethod == Methods.queryIntentActivities
-				|| mMethod == Methods.queryIntentActivityOptions || mMethod == Methods.queryIntentServices) {
+				|| mMethod == Methods.queryIntentActivityOptions || mMethod == Methods.queryIntentContentProviders
+				|| mMethod == Methods.queryIntentServices) {
 			if (param.getResult() != null && isRestricted(param))
 				param.setResult(filterResolveInfo((List<ResolveInfo>) param.getResult()));
 		} else if (mMethod == Methods.queryContentProviders) {